Refactor: index.html에서 modal-root 미리 선언 방시으로 변경
