+++ src/main.tsx
@@ -6,13 +6,6 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
 const queryClient = new QueryClient();
 
-const modalRootId = "modal-root";
-if (!document.getElementById(modalRootId)) {
-  const div = document.createElement("div");
-  div.id = modalRootId;
-  document.body.appendChild(div);
-}
-
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <QueryClientProvider client={queryClient}>
